Fixed emergency alert
