--- backend/models/Userg.js
@@ -1,10 +1,0 @@
-const mongoose = require("mongoose");
-
-const UserSchema = new mongoose.Schema({
-  name: { type: String, required: true },
-  email: { type: String, required: true, unique: true },
-  password: { type: String, required: true },
-  role: { type: String, enum: ["patient", "nurse"], required: true },
-});
-
-module.exports = mongoose.model("User", UserSchema);
