+++ backend/package.json
@@ -5,8 +5,7 @@
   "scripts": {
     "test": "echo \"Error: no test specified\" && exit 1",
     "start": "node index.js",
-    "dev": "nodemon index.js",
-    "start: graphql": "node server.js"
+    "dev": "nodemon index.js"
   }, 
   "keywords": [],
   "author": "",
