--- backend/server.js
@@ -1,20 +1,0 @@
-const app = require("./app");
-
-const express = require('express');
-const { graphqlHTTP } = require('express-graphql');
-const schema = require('./schemas');
-
-const app = express();
-
-app.use('/graphql', graphqlHTTP({
-  schema,
-  graphiql: true
-}));
-
-
-const PORT = process.env.PORT || 5000;
-
-app.listen(PORT, () => {
-  console.log(`Server running on port ${PORT}`);
-});
-
